+++ lib/init.rb
@@ -1,4 +1,6 @@
 # frozen_string_literal: true
 
-# require_relative 'core/realtime'
-# require_relative 'core/plugins'
+require_relative 'brain/redis'
+require_relative 'client/load'
+require_relative 'plugin/plugins'
+require_relative 'routes/load'
