--- startserver.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-unicorn -c $APP_HOME/unicorn.rb -D
-service nginx start
